Update README.md
Building Pan-Genomes based on Sequence Homology (BPG-SH) is a bioinformatics project that aims to construct and analyse pan-genomes of various organisms. Leveraging sequence homology, this project integrates multiple genomes to create a comprehensive representation, referred to as a pan-genome. This approach enhances our understanding of genetic diversity and evolutionary dynamics. The project uses Python and various bioinformatics tools like Clustal Omega for multiple sequence alignment. It's tailored to be highly flexible, capable of handling large genomic datasets, and designed for ease of use in genomic research